Updated and moved actions to its own folder
