+++ src/app/home/tripStats.tsx
@@ -1,5 +1,5 @@
 import styles from './page.module.css';
-import { getTripStats } from "../db";
+import { getTripStats } from "../../actions/db";
 export default async function TripStats() {
     const amountOfTrips = await getTripStats();
     return (
